Linter: Waiver file
Verilator now supports waivers for the linter to suppress known and
tolerated warnings. It is too ambitious to solve all Linter warnings
for now, but lets start with a waiver file that waivers all warnings
we see so far.

This file was generated by calling fusesoc and then cleaned up.

    fusesoc --cores-root=. run --target=lint
    chipsalliance.org:cores:SweRV_EL2 --verilator_options
    "--waiver-output verilator_waiver.vlt"

From now on the linter should warn about new warnings, and be pedantic
about it.