Github Linter Action
This adds the Github Linter Action to be invoked on every push or pull
request. It will lint the core and fail when the linter emits any
warnings not suppressed by the waiver file.

I suggest to not revert the "no-fatal" for now and be pedantic, a
linter warning should either be solved or a meaningful entry to the
waiver file added.